requirements.txt
imap_tools==1.2.0
python-dotenv==1.0.0